Changed device arrays to doubles

+++ cuda_util.cu
@@ -73,12 +73,12 @@ matrix* generate_weights(matrix const* ols_fit_matrix, matrix const* signal){
 extern "C"
 double* cuda_test_batched_ls(matrix* ls_matrix, matrix* solutions, int batch_size){
     double *A[] = {ls_matrix->data};
-    float** A_d;
+    double** A_d;
     gpu_error_check(cudaMalloc<float*>(&A_d, sizeof(A)));
     gpu_error_check(cudaMemcpy(A_d, A, sizeof(A), cudaMemcpyHostToDevice));
     
     double *C[] = {solutions->data};
-    float** C_d;
+    double** C_d;
     gpu_error_check(cudaMalloc<float*>(&C_d, sizeof(C)));
     gpu_error_check(cudaMemcpy(C_d, C, sizeof(C), cudaMemcpyHostToDevice));
